Scan 20 objects at once
This mostly addresses issue #3. There's still a chance that a single invocation might overflow the 32k Windows CLI limit, but we'll deal with that if necessary.

* Update Find-VulnerableLog4J.ps1

Replaced $FilesToScan with $ScanGroups, which splits the $Files array into an array of containing strings with 20 file path values. This causes cast.exe to scan 20 file paths at a time to prevent string limitations.